+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # gradle-nunit-plugin changelog
 ## 1.9
+### Added
 * 'labels' attribute is added to NUnit
+* Added support for multiple 'where' clauses for NUnit v3, redirected 'test' to 'where'
 
 ## 1.8
 ### Fixed